add script to make admin user before app starts
